modified the from_string routines in util to use fixed length integers as return and arguments

+++ main/lib/src/Utils.cc
@@ -90,7 +90,7 @@ namespace eudaq {
   }
 
   template<>
-    long from_string(const std::string & x, const long & def) {
+  int64_t from_string(const std::string & x, const int64_t & def) {
       if (x == "") return def;
       const char * start = x.c_str();
       char * end = 0;
@@ -102,13 +102,13 @@ namespace eudaq {
         else if (x[1] == 'x') base = 16;
         start += 2;
       }
-      long result = std::strtol(start, &end, base);
+      int64_t result = static_cast<int64_t>(std::strtol(start, &end, base));
       if (*end) throw std::invalid_argument("Invalid argument: " + x);
       return result;
     }
 
   template<>
-    uint32_t from_string(const std::string & x, const uint32_t & def) {
+    uint64_t from_string(const std::string & x, const uint64_t & def) {
       if (x == "") return def;
       const char * start = x.c_str();
       char * end = 0;
@@ -120,7 +120,7 @@ namespace eudaq {
         else if (x[1] == 'x') base = 16;
         start += 2;
       }
-      uint32_t result = std::strtoul(start, &end, base);
+      uint64_t result = static_cast<uint64_t>(std::strtoul(start, &end, base));
       if (*end) throw std::invalid_argument("Invalid argument: " + x);
       return result;
     }
